script to update sunrise and sunset
For use by other scripts, for example to only switch on lights when motion is detected at night.

This script needs a virtual device (id=90 in example) with two buttons. One to set the sun to 'Risen' and one to 'Set'. They should update the global var 'Sun'. Add some nice icons to the buttons and you'll be able to see the status of the sun easily :-)